requirements.txt
pandas==1.4.0
numpy==1.21.6
matplotlib>=3.5.1
seaborn>=0.11.2
folium>=0.12.1
ipywidgets>=7.7.0
requests==2.27.1
geopandas==0.10.2
shapely==1.8.1
PyYAML==6.0
kaggle==1.5.12 